examples/flava/requirements.txt
Pillow==9.0.1
pytorch-lightning==1.6.0
datasets==2.6.2
requests==2.27.1
DALL-E==0.1
omegaconf==2.1.2
hydra-core==1.1.2
transformers==4.16.0
pycocotools==2.0.4
